240315 Lombok, modelMapper, spring web 추가 / dtoToEntity, entityToDto 추가(CrudAuto)

+++ src/main/java/org/example/module/AddCommas.java
@@ -1,7 +1,7 @@
 package org.example.module;
 
 public class AddCommas {
-    public String addComma(int price){
+    public String addComma(Long price){
         String pToString = String.valueOf(price);
         StringBuilder result = new StringBuilder();
